wafflemarket/requirements.txt
asgiref==3.4.1
Django==3.2.6
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-jwt
django-rest-authtoken
django-filter
drf-yasg==1.20.0 # for swagger
factory-boy # for django test
django_storages # for file storage
django-google-oauth # for google login
pillow # for imagefield usage
boto3 # for S3 connection